requirements.txt
# YouTube Intelligence System - Python Dependencies

# Firebase & Core
firebase-admin>=6.4.0
python-dotenv>=1.0.1

# Analyzer (Phase 2) - Gemini AI & Image Processing
google-generativeai>=0.3.0
google-genai>=1.0.0
pydantic>=2.0.0
Pillow>=10.2.0
tqdm>=4.66.2

# Insights (Phase 3) - Statistical Analysis
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0

# Testing
pytest>=8.0.0
